refactor: better function interface, pull privkey logic out to api
