+++ pkg/license/license.go
@@ -2,27 +2,14 @@ package license
 
 import (
 	"crypto/rsa"
-	"os"
 	"time"
 
-	"github.com/ebauman/golicense/pkg/certificate"
 	golicense "github.com/ebauman/golicense/pkg/license"
 	"github.com/ebauman/golicense/pkg/types"
 	"github.com/google/uuid"
 )
 
-var PRIVATEKEY *rsa.PrivateKey
-
-func init() {
-	var GOLICENSE_KEY = os.Getenv("GOLICENSE_KEY")
-	var err error
-	PRIVATEKEY, err = certificate.PEMToPrivateKey([]byte(GOLICENSE_KEY))
-	if err != nil {
-		panic(err)
-	}
-}
-
-func CreateCarbideLicense(nodeCount int, customerID string, expirationDate time.Time) (*string, error) {
+func CreateCarbideLicense(privateKey *rsa.PrivateKey, nodeCount int, customerID string, expirationDate time.Time) (*string, error) {
 	grants := map[string]int{
 		"compliance.cattle.io/stigatron": nodeCount,
 	}
@@ -36,7 +23,7 @@ func CreateCarbideLicense(nodeCount int, customerID string, expirationDate time.
 		NotBefore: notBeforeTime,
 		Licensee:  customerID,
 	}
-	keystring, err := golicense.GenerateLicenseKey(PRIVATEKEY, license)
+	keystring, err := golicense.GenerateLicenseKey(privateKey, license)
 	if err != nil {
 		return nil, err
 	}
